Connects bot to Discord channel
Enables the bot to connect to a specific Discord channel.

The bot now interacts with the channel using the injected connection service, confirming successful or unsuccessful connection attempts. Adds `inversify` and `reflect-metadata` packages for dependency injection.

Further improvements can be made to the dependency injection to work better with discordx, but works enough for now